+++ src/test/java/com/goorm/friendchise/global/auth/application/AuthServiceTest.java
@@ -1,7 +1,11 @@
 package com.goorm.friendchise.global.auth.application;
 
+import com.goorm.friendchise.domain.auth.infrastructure.FakeRefreshTokenRepository;
 import com.goorm.friendchise.domain.manager.domain.Manager;
 import com.goorm.friendchise.domain.manager.infrastructure.FakeManagerRepository;
+import com.goorm.friendchise.global.auth.domain.RefreshTokenRepository;
+import com.goorm.friendchise.global.auth.jwt.JwtProperties;
+import com.goorm.friendchise.global.auth.jwt.TokenProvider;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -19,8 +23,10 @@ class AuthServiceTest {
 
 	@BeforeEach
 	void setUp() {
+		TokenProvider tokenProvider = new TokenProvider(new JwtProperties());
+		RefreshTokenRepository refreshTokenRepository = new FakeRefreshTokenRepository();
 		FakeManagerRepository managerRepository = new FakeManagerRepository();
-		this.authService = new AuthService(managerRepository);
+		this.authService = new AuthService(managerRepository, tokenProvider, refreshTokenRepository);
 
 		managerRepository.save(
 			Manager.create("test", "test1234", HEADQUARTER)
